Added "Replace <LayerName> with <LayerName>" command
This is pretty complicated since it is the first command that allows for two named targets in the same command string, and depends on the "with" key word sometime after "Replace" to identify the second named target. It works in basic tests but it may be more finicky than other commands when using different arrangements of words and filler words or naming layers the same as non-layer-related key words and so on. Hopefully it will be fine, since there has been multiple times I've wanted it as an option when thinking about control scheme setups.