Menus should now define a Back= command to remove the layer controlling them
This allows more control over what exactly happens when entirely back out of a menu, while also removing code just guessing which layer to actually remove, at the cost of needing to define how to close a menu manually in the Profile. Sub-menus can also define a Back= command to execute when backing to a previous menu.